requirements.txt
install==1.3.5
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
